requirements.txt
pandas
ptflops
numpy
hydra-core
pillow
requests